+++ recipes/toolkit.rb
@@ -5,7 +5,6 @@
 
 include_recipe 'percona::package_repo'
 
-unless node['percona']['version'].to_f >= 8.0 && platform_family?('rhel') &&
-       node['platform_version'].to_i >= 8
+unless node['percona']['version'].to_i >= 8 && platform_family?('rhel') && node['platform_version'].to_i >= 8
   package 'percona-toolkit'
 end
