+++ test/integration/inspec/controls/toolkit_spec.rb
@@ -4,7 +4,7 @@ control 'toolkit' do
   desc 'Ensure Percona toolkit are installed.'
   impact 1.0
 
-  if version.to_f >= 8.0 && os.family == 'redhat' && os.release.to_i >= 8
+  if version.to_i >= 8 && os.family == 'redhat' && os.release.to_i >= 8
     describe package 'percona-toolkit' do
       it { should_not be_installed }
     end
